Removed the Invoice Service and completed handling of invoicing in OrdersAPI
Invoice service was unnecesary , much easier to handle in just the order service.

The update method generates an invoice or schedules one depending on the situation. It then sends a message to the messageservice (Although I can't get a hold of the message service team currently to find out how best to handle this).